# server/swagger/swagger.yaml
openapi: 3.0.0
info:
  description: |
    UPchieve backend API
    [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net, #swagger](http://swagger.io/irc/).
  version: "1.0.0"
  title: UPchieve
  contact:
    email: tech@upchieve.org
  license:
    name: Non-Profit Open Software License 3.0
    url: 'https://gitlab.com/upchieve/subway/-/blob/main/LICENSE'
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/upchieveorg/subway/1.0.0
  - url: 'https://app.upchieve.org/v1'
...
paths:
...
  /edu:
    get:
      tags:
        - education
      summary:  Gets all courses
      description: |
        - Returns information of course categories and subcategories
        - Categories link to questions
      security:
        - cookieAuth: []
      responses:
        '304':
          description: Responds with view of categories and subcategories
  /edu/questions:
    get:
      tags:
        - education
      summary:  Gets all questions
      security:
        - cookieAuth: []
      responses:
        '304':
          description: Responds with questions view
    post:
      tags:
        - education
      summary: Creates a question in edu admin
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                questions:
                  type: object
                  properties:
                    category:
                      type: string
                    correctAnswer:
                      type: string
                    imageSrc:
                      type: string
                    questionText:
                      type: string
                    subcategory:
                      type: string
                    possibleAnswers:
                      type: array
                      items:
                        type: object
                        properties:
                          txt:
                            type: string
                          val:
                            type: string
      responses:
        '200':
          description: Returns with object of created question
...
  /edu/categoryquestions:
    post:
      tags:
        - education
      summary:  Submits details for category specific questions
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category:
                  type: string
                skip:
                  type: string
                limit:
                  type: string
      responses:
        '200':
          description: Returns object with questions of given category
  /edu/questions/{id}:
    put:
      tags:
        - education
      summary: Updates a question
      descripttion: Submits details for question update
      security:
        - cookieAuth: []
      parameters:
        - name: id
          required: true
          in: path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                questions:
                  type: object
                  properties:
                    category:
                      type: string
                    correctAnswer:
                      type: string
                    imageSrc:
                      type: string
                    questionText:
                      type: string
                    subcategory:
                      type: string
                    possibleAnswers:
                      type: array
                      items:
                        type: object
                        properties:
                          txt:
                            type: string
                          val:
                            type: string
      responses:
        '200':
          description: Returns object of updated question
    delete:
      tags:
        - education
      summary: Deletes a question from given id
      security:
        - cookieAuth: []
      parameters:
        - name: id
          required: true
          in: path
      responses:
        '200':
          description: Returns question that is deleted
...